refactor: 메인과 일치하도록 테스트 패키지 구조 변경
