+++ src/test/java/roomescape/reservation/controller/AdminReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
@@ -7,9 +7,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.reservation.controller.UserReservationController;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.dto.ReservationDeleteRequest;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.service.ReservationService;
 import roomescape.reservationtime.domain.ReservationTime;
@@ -17,15 +15,16 @@ import roomescape.theme.domain.Theme;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
 
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(UserReservationController.class)
-class UserReservationControllerTest {
+@WebMvcTest(AdminReservationController.class)
+class AdminReservationControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
@@ -37,30 +36,9 @@ class UserReservationControllerTest {
     private ReservationService reservationService;
 
     @Test
-    void 예약_목록을_조회할_수_있다() throws Exception {
+    void 관리자가_예약을_생성할_수_있다() throws Exception {
         ReservationTime time = new ReservationTime(1L, LocalTime.of(10, 0));
-        Theme theme = new Theme(2L, "Theme A", "desc", "https://example.com/a.png");
-        Reservation reservation = new Reservation(3L, "브라운", LocalDate.of(2026, 5, 1), time, theme);
-
-        when(reservationService.findAll()).thenReturn(List.of(reservation));
-
-        mockMvc.perform(get("/reservations")
-                        .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$[0].id").value(3))
-                .andExpect(jsonPath("$[0].name").value("브라운"))
-                .andExpect(jsonPath("$[0].date").value("2026-05-01"))
-                .andExpect(jsonPath("$[0].time.id").value(1))
-                .andExpect(jsonPath("$[0].time.startAt").value("10:00:00"))
-                .andExpect(jsonPath("$[0].theme.id").value(2))
-                .andExpect(jsonPath("$[0].theme.name").value("Theme A"));
-    }
-
-    @Test
-    void 예약을_생성할_수_있다() throws Exception {
-        ReservationTime time = new ReservationTime(1L, LocalTime.of(10, 0));
-        Theme theme = new Theme(2L, "Theme A", "desc", "https://example.com/a.png");
+        Theme theme = new Theme(2L, "Theme A", "desc", "thumb");
         Reservation reservation = new Reservation(3L, "브라운", LocalDate.of(2026, 5, 1), time, theme);
 
         ReservationRequest request = new ReservationRequest("브라운", LocalDate.of(2026, 5, 1), 1L, 2L);
@@ -68,7 +46,7 @@ class UserReservationControllerTest {
         when(reservationService.save(eq("브라운"), eq(LocalDate.of(2026, 5, 1)), eq(1L), eq(2L)))
                 .thenReturn(reservation);
 
-        mockMvc.perform(post("/reservations")
+        mockMvc.perform(post("/admin/reservations")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request)))
                 .andExpect(status().isCreated())
@@ -82,12 +60,8 @@ class UserReservationControllerTest {
     }
 
     @Test
-    void 예약을_삭제할_수_있다() throws Exception {
-        ReservationDeleteRequest request = new ReservationDeleteRequest("브라운");
-
-        mockMvc.perform(delete("/reservations/{id}", 1L)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request)))
+    void 관리자가_예약을_삭제할_수_있다() throws Exception {
+        mockMvc.perform(delete("/admin/reservations/{id}", 1L))
                 .andExpect(status().isNoContent());
     }
 }
